removed * imports

+++ UserService/src/main/java/com/user/userservice/controller/UserController.java
@@ -2,7 +2,18 @@ package com.user.userservice.controller;
 
 import com.nimbusds.jose.util.Pair;
 import com.user.userservice.constants.ControllerConstants;
-import com.user.userservice.dto.*;
+import com.user.userservice.dto.ApiResponse;
+import com.user.userservice.dto.UserLoginDTO;
+import com.user.userservice.dto.UserRegistrationDTO;
+import com.user.userservice.dto.UserResponseDTO;
+import com.user.userservice.dto.CountryListDTO;
+import com.user.userservice.dto.FileResponse;
+import com.user.userservice.dto.PasswordDTO;
+import com.user.userservice.dto.UserUpdateDTO;
+import com.user.userservice.dto.UpdateRecipeDTO;
+import com.user.userservice.dto.UserEmailDTO;
+import com.user.userservice.dto.AddRecipeDTO;
+import com.user.userservice.dto.UserDisplayDTO;
 import com.user.userservice.feignclient.RecipeServiceClient;
 import com.user.userservice.entity.User;
 import com.user.userservice.exception.IncorrectPasswordException;
@@ -153,7 +164,11 @@ public class UserController {
     @PutMapping("recipes/delete/{id}")
     public ResponseEntity<ApiResponse> deleteRecipe(@PathVariable("id") Long id,@RequestParam("userid")Long userId) throws InvalidInputException, UserNotFoundException {
         UserDisplayDTO user=userService.getUser(id);
-        String s= recipeServiceClient.getRecipeOwnerId(id).getBody();
+        ResponseEntity<String> ownerResponse = recipeServiceClient.getRecipeOwnerId(id);
+        if (ownerResponse == null || ownerResponse.getBody() == null) {
+            throw new NullPointerException("Failed to retrieve owner information for recipe ID: " + id);
+        }
+        String s= ownerResponse.getBody();
         if(s.compareTo(String.valueOf(userId))==0){
             recipeServiceClient.deleteRecipe(id);
             ApiResponse response = ApiResponse.builder()
